Fixed a bug in rush mode closing cards which shouldn't.
This happened when somebody (A) opens a card, then another player (B)
opens this same card and its corresponding pair, thus making them
open for everyone. Then when A clicks somewhere else to, the pair
is wrong and it closes again even though it should have stood open
because B opened it for all.