Server-side relative card-center coordinates.
This means that on the server the card coordinates are between 0 and 1,
the server has no concept of either card or board size. He distributes
the cards on a grid (plus some randomness soon).

The client keeps the card coordinates as absolute pixels between 0 and
boardsize. Probably the board size could depend on the window size or
screen resolution and the card sizes on the assets.